+++ libs/common/src/lib/resources/edit-resource/edit-resource.module.ts
@@ -15,14 +15,16 @@ import {
   ResourceTextComponent,
   SelectFirstFieldDirective,
 } from './';
-import { DropdownButtonComponent, SisProgressModule } from '@nuclia/sistema';
+import { DropdownButtonComponent, JsonViewerComponent, SisProgressModule } from '@nuclia/sistema';
 import {
   PaButtonModule,
   PaChipsModule,
+  PaDatePickerModule,
   PaDropdownModule,
   PaExpanderModule,
   PaIconModule,
   PaScrollModule,
+  PaTableModule,
   PaTextFieldModule,
   PaTogglesModule,
   PaTooltipModule,
@@ -59,6 +61,9 @@ import { HintModule } from '../../hint';
 
     STFPipesModule,
     PaTooltipModule,
+    PaDatePickerModule,
+    PaTableModule,
+    JsonViewerComponent,
   ],
   declarations: [
     AddFieldComponent,