+++ libs/common/src/lib/resources/edit-resource/profile/profile.component.ts
@@ -7,6 +7,7 @@ import { filter, forkJoin, map, Observable, Subject, switchMap, tap, timer } fro
 import { SisToastService } from '@nuclia/sistema';
 import { takeUntil } from 'rxjs/operators';
 import { EditResourceService } from '../edit-resource.service';
+import { JsonValidator } from '../../../validators';
 
 type Thumbnail = { uri: string; blob: SafeUrl };
 
@@ -41,6 +42,10 @@ export class ResourceProfileComponent implements OnInit {
       modified: new FormControl<string>('', { nonNullable: true }),
       related: new FormControl<string>('', { nonNullable: true }),
     }),
+    extra: new FormControl<string>('', {
+      nonNullable: true,
+      validators: [JsonValidator()],
+    }),
   });
   thumbnails: Observable<Thumbnail[]> = this.resource.pipe(
     map((res) => this.getThumbnailsAndImages(res)),
@@ -68,6 +73,8 @@ export class ResourceProfileComponent implements OnInit {
 
   isUploading = false;
   isSaving = false;
+  extraMetadata: any;
+  editExtraMetadata = false;
 
   constructor(
     private editResource: EditResourceService,
@@ -102,7 +109,9 @@ export class ResourceProfileComponent implements OnInit {
         modified: data.origin?.modified,
         related: (data.origin?.related || []).join('\n'),
       },
+      extra: JSON.stringify(data.extra?.metadata, null, 2),
     });
+    this.extraMetadata = data.extra?.metadata;
     this.form.enable();
     this.cdr?.markForCheck();
   }
@@ -112,7 +121,9 @@ export class ResourceProfileComponent implements OnInit {
     const data: Partial<Resource> = this.getValue();
     this.editResource.savePartialResource(data).subscribe(() => {
       this.form.markAsPristine();
+      this.extraMetadata = data.extra?.metadata;
       this.isSaving = false;
+      this.editExtraMetadata = false;
       this.cdr.markForCheck();
     });
   }
@@ -136,10 +147,11 @@ export class ResourceProfileComponent implements OnInit {
             collaborators: value.origin.collaborators.split(',').map((s) => s.trim()),
             url: value.origin.url,
             filename: value.origin.filename,
-            created: value.origin.created ? new Date(value.origin.created).toISOString() : undefined,
-            modified: value.origin.modified ? new Date(value.origin.modified).toISOString() : undefined,
+            created: value.origin.created || undefined,
+            modified: value.origin.modified || undefined,
             related: value.origin.related.split('\n').map((s) => s.trim()),
           },
+          extra: value.extra ? { metadata: JSON.parse(value.extra) } : undefined,
         }
       : {};
   }